add broken hw7 for willghatch
Sorry, I've been really busy with other things and haven't given this
assignment the time it deserves.  I'll try to set aside some time to
debug things and make them work...

Before starting I looked at the solutions that were turned in to see
what files needed to be changed.  But clearly I didn't make as good of
a solution...